requirements.txt
aiohttp==3.8.1
anyio==3.6.2
asyncer==0.0.1
click==8.1.3
fastapi==0.85.1
filetype==1.1.0
Flask==2.1.2
gdown==4.5.1
numpy==1.21.6
onnxruntime==1.13.1
onnxruntime_gpu==1.12.1
opencv_python_headless==4.6.0.66
Pillow==9.2.0
PyMatting==1.1.8
scipy==1.7.3
starlette==0.19.1
tqdm==4.64.0
uvicorn==0.19.0
watchdog==2.1.9
Werkzeug==2.1.2
